Pick ConfigParser changes from apache@0c3614e
