+++ cordova-common/spec/fixtures/test-config.xml
@@ -9,6 +9,71 @@
     </author>
     <content src="index.html" />
     <access origin="*" />
+
+    <!--
+        <access> tests
+
+        Test wildcard allow all. Note in ATS you can open up all, and have restrictions for certain domains.
+        This is to allow for example, an in-app-browser (*) but your own server communications must be https, for example.
+     -->
+    <access origin="*" />
+    <!-- https, with subdomain wildcard, with attribute differences -->
+    <access origin="https://*.server01.com" /> <!-- equivalent to the next line -->
+    <access origin="https://*.server01.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <access origin="https://*.server02.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <access origin="https://*.server03.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <access origin="https://*.server04.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+    <!-- http, no subdomain wildcard, with attribute differences -->
+    <access origin="http://server05.com" /> <!-- equivalent to the next line -->
+    <access origin="http://server05.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <access origin="http://server06.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <access origin="http://server07.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <access origin="http://server08.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+    <!-- https, no subdomain wildcard, with attribute differences -->
+    <access origin="https://server09.com" /> <!-- equivalent to the next line -->
+    <access origin="https://server09.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <access origin="https://server10.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <access origin="https://server11.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <access origin="https://server12.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+
+    <!--
+        <allow-navigation> tests
+
+        Test wildcard allow all. Note in ATS you can open up all, and have restrictions for certain domains.
+        This is to allow for example, an in-app-browser (*) but your own server communications must be https, for example.
+     -->
+    <allow-navigation href="*" />
+    <!-- https, with subdomain wildcard, attribute differences -->
+    <allow-navigation href="https://*.server21.com" /> <!-- equivalent to the next line -->
+    <allow-navigation href="https://*.server21.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <allow-navigation href="https://*.server22.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <allow-navigation href="https://*.server23.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <allow-navigation href="https://*.server24.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+    <!-- http, no subdomain, with attribute differences -->
+    <allow-navigation href="http://server25.com" /> <!-- equivalent to the next line -->
+    <allow-navigation href="http://server25.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <allow-navigation href="http://server26.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <allow-navigation href="http://server27.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <allow-navigation href="http://server28.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+    <!-- https, no subdomain, with attribute differences -->
+    <allow-navigation href="https://server29.com" /> <!-- equivalent to the next line -->
+    <allow-navigation href="https://server29.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <allow-navigation href="https://server30.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <allow-navigation href="https://server31.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <allow-navigation href="https://server32.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+    <!-- wildcard scheme, with subdomain wildcard, with attribute differences -->
+    <allow-navigation href="*://*.server33.com" /> <!-- equivalent to the next line -->
+    <allow-navigation href="*://*.server33.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <allow-navigation href="*://*.server34.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <allow-navigation href="*://*.server35.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <allow-navigation href="*://*.server36.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+    <!-- wildcard scheme, no subdomain, with attribute differences -->
+    <allow-navigation href="*://server37.com" /> <!-- equivalent to the next line -->
+    <allow-navigation href="*://server37.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="true" />
+    <allow-navigation href="*://server38.com" minimum-tls-version="TLSv1.2" requires-forward-secrecy="false" />
+    <allow-navigation href="*://server39.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="true" />
+    <allow-navigation href="*://server40.com" minimum-tls-version="TLSv1.1" requires-forward-secrecy="false" />
+
     <preference name="fullscreen" value="true" />
     <preference name="webviewbounce" value="true" />
     <preference name="orientation" value="portrait" />
